Flags set correctly on appropriate instructions
Flags should set correctly for all of the instructions now. I changed ov
to msb because only add and subtract should set the ov (and ne) flags,
but ov was simply being pulled from the result of the alu - even on
something like a shift or AND, where overflow isn't possible.

Also added an input to the alu letting it know if it is a typical aluOp
or not (again so it knows when to set the zero flag)